requirements.txt
boto3
python-dotenv
ipdb
PyPDF2
python-docx
strands-agents
strands-agents-tools 
strands-agents-builder
strands-agents[openai]
